create trajectory, point_helper class
