+++ obstacle_stop_planner_refine/include/obstacle_stop_planner/vehicle.hpp
@@ -84,6 +84,6 @@ public:
 private:
 };
 
-}
+}  // namespace motion_planning
 
 #endif  // OBSTACLE_STOP_PLANNER__VEHICLE_HPP_
